Architecture: Split dashboard controller into packages  + testing
